- Flash invocation callback handle with a better behavior on multiple response on one requestHandle - A HTTP secure issue fixed on a bad HTTP requester - Better understanding of Flash timestamp, and H264/AAC infos packet give a major refactorization of Publication.cpp and Listener.cpp class - Mona is compatible with Flash Media Live Encoder now - Writer::hasToConvert removing and replaced by dynamic code - Session::onAddressChange event addition to replace the old bad design when client address changeWriter - A array bug in AMFWriter fixed - Event::lock function addition to lock the event if required - FileSystem::Remove with optional argument "all" to indicate that it have to remove all recursivly (more safe like that with before where it was the default behavior) - a WebSocket infinite loop fixed on client death - a RTMFP publication, data could be lost, fixed - endianness problem fixed in RTMFP.h, String.h and StringTest.cpp - a memory leak issue fixed in IPAddress.cpp and SocketAddress.cpp - a linux starting issue fixed - Minor inclusion, logs level and function signature change

+++ MonaCore/sources/RTMFP/RTMFPSession.cpp
@@ -276,7 +276,7 @@ void RTMFPSession::packetHandler(PacketReader& packet) {
 				time += 0xFFFF-timeEcho;
 			timeEcho = 0;
 		}
-		peer.setPing((time-timeEcho)*(UInt16)RTMFP_TIMESTAMP_SCALE);
+		peer.setPing((time-timeEcho)*RTMFP_TIMESTAMP_SCALE);
 	}
 	else if(marker != 0xF9)
 		WARN("RTMFPPacket marker unknown : ", Format<UInt8>("%02x",marker));
